test completed - aggregation
