+++ .pre-commit-config.yaml
@@ -13,7 +13,7 @@ repos:
     hooks:
       - id: black
         # setup.cfgに書きたいがpre commitが読んでくれない
-        exclude: ^pnq_template/template.py
+        exclude: ^pnq/__template__.py
   # - repo: https://github.com/pre-commit/mirrors-mypy
   #   rev: "v0.700"
   #   hooks:
