+++ pnq/exceptions.py
@@ -3,8 +3,8 @@ class PnqException(Exception):
     Pnqに関連する全ての例外の基底クラス
     """
 
-    def __init__(self):
-        pass
+    def __init__(self, msg: str = ""):
+        super().__init__(msg)
 
 
 # TODO: おそらくpnq in pnqでStopIterationが暴露した時に、想定外の事象が発生する
@@ -12,8 +12,16 @@ class StopIteration(PnqException):
     pass
 
 
+class MustError(PnqException):
+    """
+    クエリがシーケンス全体あるいはシーケンスの要素に対して任意の特性を持っていることを要求したが失敗した
+
+    関連: `must_unique`
+    """
+
+
 # KeyError IndexErrorを継承しても try: ... except KeyError としても補足できないっぽい
-class KeyNotFoundError(PnqException):
+class KeyNotFoundError(MustError):
     """
     クエリがキーに対応する要素を要求したが存在しない。
     IndexErrorとKeyErrorはKeyNotFoundErrorに置き換わります
@@ -24,7 +32,7 @@ class KeyNotFoundError(PnqException):
     pass
 
 
-class NoElementError(PnqException):
+class NoElementError(MustError):
     """
     クエリが何らかの要素を要求したが要素が存在しない
 
@@ -34,7 +42,7 @@ class NoElementError(PnqException):
     pass
 
 
-class NotOneElementError(PnqException):
+class NotOneElementError(MustError):
     """
     クエリが要素がひとつであることを要求したが複数の要素が存在した
 
@@ -44,7 +52,7 @@ class NotOneElementError(PnqException):
     pass
 
 
-class DuplicateElementError(PnqException):
+class DuplicateElementError(MustError):
     """
     クエリが要素が重複していないことを要求したが重複を検知した
